refactor app and app_state modules into separate folders/files
