+++ src/app_state/app.rs
@@ -1,0 +1,42 @@
+use std::path::PathBuf;
+use std::collections::HashMap;
+use std::fs;
+
+use super::ui::{FileTreeEntry, UIState};
+
+pub struct AppState {
+    /// Directory of the entire project, can only be a single one
+    /// It is either opened directly (like `love .`) or it is calculated
+    /// based on the provided path
+    pub working_directory: PathBuf,
+    pub file_tree: HashMap<PathBuf, Vec<FileTreeEntry>>,
+    pub lines: Vec<Vec<char>>,
+    pub ui_state: UIState,
+}
+
+impl AppState {
+    pub fn new(file_content: String, working_directory: PathBuf) -> AppState {
+        let lines_number = file_content.lines().count();
+        let lines: Vec<Vec<char>> = file_content.lines().map(|s| s.chars().collect()).collect();
+
+        AppState {
+            working_directory,
+            file_tree: HashMap::new(),
+            lines,
+            // this is extremely safe, it needs to have 65535 digits to overflow
+            ui_state: UIState::new(lines_number.to_string().len() as u16)
+        }
+    }
+
+    pub fn read_directory(&mut self, path: PathBuf) {
+        // TODO: save the error as a directory state
+        let dir_entries = fs::read_dir(&path).expect("Could not read directory");
+
+        let values: Vec<FileTreeEntry> = dir_entries
+            .filter_map(|result| result.ok())
+            .map(|entry| FileTreeEntry::new(entry.path()))
+            .collect();
+
+        self.file_tree.insert(path, values);
+    }
+}
